backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
aioredis==2.0.1
pyjwt==2.9.0
python-dotenv==1.0.1

